Skills/TeachingAffairs/pdf-image-extractor/requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0
PyYAML>=6.0
Jinja2>=3.1.0
tqdm>=4.66.0
